Implement XMODEM file transfer over BLE/UART
Added an XMODEM receiver implementation to the mbed app. This also
required the extension of the builtin UART service implementation
and the replacement of the previous CStringBuffer implementation
with a generic byte Buffer. If a transfer is successful, the app
prints the contents to the USB serial line.

Also added a Python tool that acted as an XMODEM sender over
BLE/UART. The sender implementation is based on PyPI-hosted
pip-installable packages.