JS execution in separate thread
The "init JS engine" / "parse & execute JS code" / "engine shutdown"
steps are now performed in a never-ending loop in a separate thread
that starts a new iteration every time a new JS file is uploaded via
BLE/UART/XMODEM. To enable this, the engine is now compiled with
`JERRY_VM_EXEC_STOP`, so that it can be stopped and shut down even
if the JS code is running infinitely.

Also added an `examples` directory to contain sample JS snippets
that can be uploaded to the device.

Related refactorings:
- Modularized the code: moved {Java,Jerry}Script-related logic to
  `jrs-*.{h,cpp}`, and Bluetooth LE-related logic to `blue.{h,cpp}`.
- Applied consistent naming scheme (`jrs_` and `blue_` prefixes in
  identifiers, and `[jrs]` and `[blue]` prefixes in log messages
  everywhere in the new sources, and `[main]` in the remaining
  `main.cpp`).
- Merged `peripherals.{h,cpp}` back into `main.{h,cpp}`.