Add support for Current Time Service to the HTML file upload tool
Now, when a script is uploaded, txjs.html can set the RTC, too.
This way the Python and the HTML tool variants got the same
functionality.

To enable adding the CTS support, the modularization of the JS code
had to be reworked. This affected the responsibilities and the API
of the UART module, as well as the tasks of the main txjs module.